backend/microservice-STT/requirements.txt
fastapi
uvicorn
httpx
aioboto3
pydantic
google-genai 